git/ci: Replace tj-actions/changed-files.
This repo has been deleted after it was compromised. We do not have
any secrets in our repo and we had also pinned the version of the
action to sha so there is no issue for us.

But the repo is gone now. I've instead matched this code to what we
are doing in the gitlab CI runner where we did not have such an
action anyway.

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>